fxml для главного окна
