+++ src/ru/avalon/javapp/devj140/userGUI/MainApplication/FXMLUserGUI.fxml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<?import javafx.scene.text.Text?>
+<GridPane fx:id="root" xmlns:fx="http://javafx.com/fxml/1" hgap="25" vgap="15" style="-fx-padding:20"
+          alignment="CENTER" prefHeight="300.0" prefWidth="500.0"
+          xmlns="http://javafx.com/javafx/"
+          fx:controller="ru.avalon.javapp.devj140.userGUI.MainApplication.FXMLUserGUIController">
+
+    <children>
+        <Label text="Welcom" GridPane.columnIndex="0" GridPane.rowIndex="0" GridPane.columnSpan="2" style="-fx-font-size:20"/>
+        <Label text="User Name" GridPane.columnIndex="0" GridPane.rowIndex="1"/>
+        <TextField
+                fx:id="userNameText"
+                GridPane.columnIndex="1"
+                GridPane.rowIndex="1"
+                onKeyPressed="#clearMessage"
+        />
+        <Label text="User Password" GridPane.columnIndex="0" GridPane.rowIndex="2"/>
+        <TextField
+                fx:id="userPasswordText"
+                GridPane.columnIndex="1"
+                GridPane.rowIndex="2"
+                onKeyPressed="#clearMessage"
+        />
+        <Button
+                fx:id="signIn"
+                text="Sign in"
+                GridPane.columnIndex="1"
+                GridPane.rowIndex="3"
+                GridPane.halignment="RIGHT"
+                onAction="#checkUser"
+        />
+        <Text fx:id="message" GridPane.columnIndex="0" GridPane.columnSpan="2" GridPane.rowIndex="5" style="-fx-font-size:15"/>
+    </children>
+
+</GridPane>
