+++ src/ru/avalon/javapp/devj140/userGUI/MainApplication/Message.java
@@ -1,0 +1,7 @@
+package ru.avalon.javapp.devj140.userGUI.MainApplication;
+
+public enum Message {
+    ERROR,
+    WARNING,
+    OK
+}
